refactor: rename link module to key module
- Rename modules/libs/src/model/link.rs to key.rs
- Update module declaration in model.rs
